Changed MatchState to CompetitionState Potential Fix for mob spawns and owners

+++ src/java/mc/alk/arena/Defaults.java
@@ -38,7 +38,7 @@ public class Defaults {
 
 	public static boolean MATCH_FORCESTART_ENABLED = true;
 	public static boolean MATCH_FORCESTART_START_ONJOIN = true;
-	public static long MATCH_FORCESTART_TIME = 60;
+	public static int MATCH_FORCESTART_TIME = 60;
 
 	public static boolean USE_SCOREBOARD = true;
 	public static boolean USE_COLORNAMES = true;
